requirements.txt
lark-parser
pytest
oyaml
pyinstaller
